+++ src/assets/svgs/TabBarNav/index.ts
@@ -1,11 +1,11 @@
 import Exposure from "./exposure"
 import Heartbeat from "./heartbeat"
 import House from "./house"
-import PersonInHand from "./personInHand"
+import Gear from "./gear"
 
 export const TabBarIcons = {
   Exposure,
   Heartbeat,
   House,
-  PersonInHand,
+  Gear,
 }
