--- src/assets/svgs/TabBarNav/personInHand.ts
@@ -1,6 +1,0 @@
-export default `
-<svg width="27" height="27" viewBox="0 0 27 27" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M26.9981 13.4408C26.9554 11.5725 25.4233 10.0654 23.5446 10.0654C21.666 10.0654 20.1344 11.5727 20.0914 13.4408C20.0903 13.4604 20.0895 13.4797 20.0895 13.499C20.0895 15.1367 19.5064 16.3147 18.3067 17.1009C17.8396 17.4071 17.2739 17.6524 16.6243 17.8337C16.755 17.4735 16.8269 17.0851 16.8269 16.6797C16.8269 14.8283 15.3332 13.3199 13.4875 13.294C13.4755 13.2935 13.4635 13.2935 13.4513 13.2932C13.4472 13.2932 13.4436 13.293 13.4398 13.293H0.926932C0.415146 13.293 0 13.7081 0 14.2199V15.1092C0 16.9324 0.362062 18.6272 1.07611 20.146C1.75859 21.5975 2.74922 22.8639 4.02025 23.9106C6.44034 25.9025 9.80697 26.9993 13.5 26.9993C17.1062 26.9993 20.4962 25.5949 23.0459 23.0452C25.5861 20.5051 26.9888 17.1308 26.9995 13.5401C26.9997 13.5336 27 13.5273 27 13.5208C27 13.5184 26.9997 13.5162 26.9997 13.514C26.9997 13.5091 27 13.5045 27 13.4996C27 13.4797 26.9989 13.4604 26.9981 13.4408ZM13.5 25.1452C7.71409 25.1452 1.87646 22.0532 1.85495 15.1471H13.4401C14.2854 15.1471 14.973 15.8347 14.973 16.6797C14.973 17.5253 14.2856 18.2126 13.4401 18.2126H7.32644C6.81465 18.2126 6.39951 18.6278 6.39951 19.1398C6.39951 19.6519 6.81465 20.0668 7.32644 20.0668H13.6233C13.6364 20.0668 13.65 20.0665 13.6631 20.0657C16.0184 20.045 17.9221 19.5697 19.3232 18.6517C21.0268 17.5353 21.932 15.7689 21.9434 13.5415C21.9434 13.5347 21.9437 13.5273 21.9437 13.5205C21.9437 12.6374 22.6621 11.9196 23.5446 11.9196C24.425 11.9196 25.1418 12.6342 25.1453 13.5132C25.138 19.9282 19.9169 25.1452 13.5 25.1452Z" />
-<path d="M7.60494 12.8777C8.11673 12.8777 8.53187 12.4631 8.53187 11.9508C8.53187 9.53883 10.4941 7.57689 12.9057 7.57689C15.3177 7.57689 17.2799 9.5391 17.2799 11.9508V14.6654C17.2799 15.1775 17.6948 15.5926 18.2068 15.5926C18.7186 15.5926 19.1337 15.1775 19.1337 14.6654V11.9508C19.1337 9.48193 17.6901 7.34441 15.6024 6.33717C16.2503 5.66395 16.6491 4.74927 16.6491 3.74339C16.6491 1.6791 14.9698 0 12.9057 0C10.8417 0 9.16208 1.6791 9.16208 3.74339C9.16208 4.74954 9.56116 5.66395 10.2091 6.33717C8.12163 7.34414 6.67773 9.48193 6.67773 11.9508C6.67773 12.4628 7.09261 12.8777 7.60494 12.8777ZM12.906 1.85386C13.9481 1.85386 14.7955 2.70158 14.7955 3.74312C14.7955 4.7852 13.9481 5.63265 12.906 5.63265C11.8639 5.63265 11.0165 4.78493 11.0165 3.74312C11.0162 2.70158 11.8639 1.85386 12.906 1.85386Z" />
-</svg>
-`
