--- src/assets/svgs/gear.ts
@@ -1,7 +1,0 @@
-export default `
-<svg width="26" height="26" viewBox="0 0 26 26" xmlns="http://www.w3.org/2000/svg">
-<path d="M13.8163 26H12.1837C11.307 26 10.5814 25.2744 10.5814 24.3977V23.4302C10.5814 23.2488 10.4605 23.0977 10.2791 23.0674C9.40233 22.8256 8.58605 22.493 7.8 22.0395C7.64884 21.9488 7.43721 21.9791 7.31628 22.1L6.65116 22.7651C6.01628 23.4 5.0186 23.4 4.38372 22.7651L3.23488 21.6163C2.6 20.9814 2.6 19.9837 3.23488 19.3488L3.9 18.6837C4.02093 18.5628 4.05116 18.3814 3.96047 18.2C3.50698 17.414 3.17442 16.5674 2.93256 15.7209C2.87209 15.5395 2.72093 15.4186 2.56977 15.4186H1.60233C0.725581 15.4186 0 14.693 0 13.8163V12.1837C0 11.307 0.725581 10.5814 1.60233 10.5814H2.56977C2.75116 10.5814 2.90233 10.4605 2.93256 10.2791C3.17442 9.40233 3.50698 8.58605 3.96047 7.8C4.05116 7.64884 4.02093 7.43721 3.9 7.31628L3.23488 6.65116C2.6 6.01628 2.6 5.0186 3.23488 4.38372L4.38372 3.23488C5.0186 2.6 6.01628 2.6 6.65116 3.23488L7.31628 3.9C7.43721 4.02093 7.6186 4.05116 7.8 3.96047C8.58605 3.50698 9.43256 3.17442 10.2791 2.93256C10.4605 2.87209 10.5814 2.72093 10.5814 2.56977V1.60233C10.5814 0.725581 11.307 0 12.1837 0H13.8163C14.693 0 15.4186 0.725581 15.4186 1.60233V2.56977C15.4186 2.75116 15.5395 2.90233 15.7209 2.93256C16.5977 3.17442 17.414 3.50698 18.2 3.96047C18.3512 4.05116 18.5628 4.02093 18.6837 3.9L19.3488 3.23488C19.9837 2.6 20.9814 2.6 21.6163 3.23488L22.7651 4.38372C23.4 5.0186 23.4 6.01628 22.7651 6.65116L22.1 7.31628C21.9791 7.43721 21.9488 7.6186 22.0395 7.8C22.493 8.58605 22.8256 9.43256 23.0674 10.2791C23.1279 10.4605 23.2791 10.5814 23.4302 10.5814H24.3977C25.2744 10.5814 26 11.307 26 12.1837V13.8163C26 14.693 25.2744 15.4186 24.3977 15.4186H23.4302C23.2488 15.4186 23.0977 15.5395 23.0674 15.7209C22.8256 16.5977 22.493 17.414 22.0395 18.2C21.9488 18.3512 21.9791 18.5628 22.1 18.6837L22.7651 19.3488C23.4 19.9837 23.4 20.9814 22.7651 21.6163L21.6163 22.7651C20.9814 23.4 19.9837 23.4 19.3488 22.7651L18.6837 22.1C18.5628 21.9791 18.3814 21.9488 18.2 22.0395C17.414 22.493 16.5674 22.8256 15.7209 23.0674C15.5395 23.1279 15.4186 23.2791 15.4186 23.4302V24.3977C15.4186 25.2744 14.693 26 13.8163 26ZM7.58837 20.7698C7.86046 20.7698 8.13256 20.8302 8.40465 20.9814C9.1 21.3744 9.82558 21.6767 10.6116 21.8884C11.307 22.0698 11.7907 22.7047 11.7907 23.4302V24.3977C11.7907 24.6093 11.9721 24.7907 12.1837 24.7907H13.8163C14.0279 24.7907 14.2093 24.6093 14.2093 24.3977V23.4302C14.2093 22.7047 14.693 22.0698 15.3884 21.8884C16.1442 21.6767 16.9 21.3744 17.5953 20.9814C18.2302 20.6186 19.0163 20.7093 19.5302 21.2233L20.1953 21.8884C20.3465 22.0395 20.5884 22.0395 20.7395 21.8884L21.8884 20.7395C22.0395 20.5884 22.0395 20.3465 21.8884 20.1953L21.2233 19.5302C20.7093 19.0163 20.6186 18.2302 20.9814 17.5953C21.3744 16.9 21.6767 16.1744 21.8884 15.3884C22.0698 14.693 22.7047 14.2093 23.4302 14.2093H24.3977C24.6093 14.2093 24.7907 14.0279 24.7907 13.8163V12.1837C24.7907 11.9721 24.6093 11.7907 24.3977 11.7907H23.4302C22.7047 11.7907 22.0698 11.307 21.8884 10.6116C21.6767 9.85581 21.3744 9.1 20.9814 8.40465C20.6186 7.76977 20.7093 6.98372 21.2233 6.46977L21.8884 5.80465C22.0395 5.65349 22.0395 5.41163 21.8884 5.26047L20.7395 4.11163C20.5884 3.96047 20.3465 3.96047 20.1953 4.11163L19.5302 4.77674C19.0163 5.2907 18.2302 5.3814 17.5953 5.0186C16.9 4.62558 16.1744 4.32326 15.3884 4.11163C14.693 3.93023 14.2093 3.29535 14.2093 2.56977V1.60233C14.2093 1.3907 14.0279 1.2093 13.8163 1.2093H12.1837C11.9721 1.2093 11.7907 1.3907 11.7907 1.60233V2.56977C11.7907 3.29535 11.307 3.93023 10.6116 4.11163C9.85581 4.32326 9.1 4.62558 8.40465 5.0186C7.76977 5.3814 6.98372 5.2907 6.46977 4.77674L5.80465 4.11163C5.65349 3.96047 5.41163 3.96047 5.26047 4.11163L4.11163 5.26047C3.96047 5.41163 3.96047 5.65349 4.11163 5.80465L4.77674 6.46977C5.2907 6.98372 5.3814 7.76977 5.0186 8.40465C4.62558 9.1 4.32326 9.82558 4.11163 10.6116C3.93023 11.307 3.29535 11.7907 2.56977 11.7907H1.60233C1.3907 11.7907 1.2093 11.9721 1.2093 12.1837V13.8163C1.2093 14.0279 1.3907 14.2093 1.60233 14.2093H2.56977C3.29535 14.2093 3.93023 14.693 4.11163 15.3884C4.32326 16.1442 4.62558 16.9 5.0186 17.5953C5.3814 18.2302 5.2907 19.0163 4.77674 19.5302L4.11163 20.1953C3.96047 20.3465 3.96047 20.5884 4.11163 20.7395L5.26047 21.8884C5.41163 22.0395 5.65349 22.0395 5.80465 21.8884L6.46977 21.2233C6.77209 20.9209 7.19535 20.7698 7.58837 20.7698Z" />
-<path d="M13.0001 17.9885C10.2489 17.9885 8.01172 15.7513 8.01172 13.0001C8.01172 10.2489 10.2489 8.01172 13.0001 8.01172C15.7513 8.01172 17.9885 10.2489 17.9885 13.0001C17.9885 15.7513 15.7513 17.9885 13.0001 17.9885ZM13.0001 9.22102C10.914 9.22102 9.22102 10.914 9.22102 13.0001C9.22102 15.0861 10.914 16.7792 13.0001 16.7792C15.0861 16.7792 16.7792 15.0861 16.7792 13.0001C16.7792 10.914 15.0861 9.22102 13.0001 9.22102Z" />
-</svg>
-
-`
